Added Barebones JavaScript Guidelines to Repo
Although it needs fleshing out, I've added a barebones copy of the JavaScript
guidelines to the repo and linked it from the readme.md file to make it accessable.

We'll slowly expand on it over time, but at the moment it's very basic and the
format is based on the SCSS file.